fix&feats: Change decoding and fix visualization
MAJOR UPDATE!
- Deprecate 2's complement decoding
- Implement resolution: decode more points within a fixed boundary rather than range scaling with more points
- Implement logarithmic scaling of heatmap
- Implement contour lines for plot
- Implement logarithmic scaling contour lines

+++ a3/src/dt/individual.py
@@ -2,6 +2,6 @@ from typing import TypedDict
 
 class individual(TypedDict):
     chrom: str          # The encoded string
-    x: int              # The first decoded value
-    y: int              # The second decoded value
+    x: float            # The first decoded value
+    y: float            # The second decoded value
     fitness: float      # The fitness score
